backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.1.0
PyJWT==2.10.1
django-filter==23.4
drf-spectacular==0.26.5
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1
PyPDF2
openai==1.57.4
python-docx==1.2.0
google-genai==1.62.0